use c types for the input_event and timeval structs - add serde feature that derives Serialize and Deserialize for input_event and timeval - restructure project into workspace
