+++ Cargo.toml
@@ -1,8 +1,3 @@
-[package]
-name = "carbide"
-version = "0.1.0"
-edition = "2024"
-
-[[bin]]
-name = "crb"
-path = "src/main.rs"
+[workspace]
+resolver = "3"
+members = ["carbide", "carbide_cli"]
