+++ carbide/src/main.rs
@@ -1,0 +1,9 @@
+mod cli;
+mod devices;
+mod input_event;
+
+use cli::cli;
+
+fn main() {
+    cli();
+}
